requirements-test.txt
pytest==7.0.1
pytest-django==4.5.0
djangorestframework==3.14.0